==> 10_code/11_confirm_states.py <==
"""checks states for sanity check and FL counties for discrepancies"""
import pandas as pd

csv_url = "https://raw.githubusercontent.com/wpinvestigative/arcos-api/master/data/county_fips.csv?raw=true"

df = pd.read_csv(csv_url)

# Get the unique states in the "BUYER_STATE" column
unique_states = df["BUYER_STATE"].unique()

# Self-made list of states
STATES = [
    "AK",
    "AL",
    "AZ",
    "AR",
    "CA",
    "CO",
    "CT",
    "DE",
    "FL",
    "GA",
    "HI",
    "ID",
    "IL",
    "IN",
    "IA",
    "KS",
    "KY",
    "LA",
    "ME",
    "MD",
    "MA",
    "MI",
    "MN",
    "MS",
    "MO",
    "MT",
    "NE",
    "NV",
    "NH",
    "NJ",
    "NM",
    "NY",
    "NC",
    "ND",
    "OH",
    "OK",
    "OR",
    "PA",
    "RI",
    "SC",
    "SD",
    "TN",
    "TX",
    "UT",
    "VT",
    "VA",
    "WA",
    "WV",
    "WI",
    "WY",
    "DC",
]


# Print the unique states
print(unique_states)

# Find missing states
missing_states = set(unique_states) - set(STATES)

# Print the missing states
print("Missing States:", missing_states)

check2 = {
    "ID",
    "PR",
    "NC",
    "AL",
    "MT",
    "OR",
    "MN",
    "FL",
    "KY",
    "MI",
    "SC",
    "WI",
    "WA",
    "AE",
    "NE",
    "MP",
    "WY",
    "AR",
    "DE",
    "AK",
    "LA",
    "IA",
    "TX",
    "KS",
    "NJ",
    "ME",
    "NY",
    "WV",
    "IL",
    "MD",
    "MO",
    "SD",
    "OH",
    "AZ",
    "NH",
    "CT",
    "ND",
    "PA",
    "CA",
    "MA",
    "RI",
    "NV",
    "CO",
    "VI",
    "HI",
    "PW",
    "OK",
    "DC",
    "TN",
    "MS",
    "VT",
    "NM",
    "VA",
    "IN",
    "GA",
    "GU",
    "UT",
}

missing_states = set(check2) - set(STATES)
print("Missing States:", missing_states)

fl_counties = [
    "alachua",
    "baker",
    "bay",
    "bradford",
    "brevard",
    "broward",
    "calhoun",
    "charlotte",
    "citrus",
    "clay",
    "collier",
    "columbia",
    "desoto",
    "dixie",
    "duval",
    "escambia",
    "flagler",
    "franklin",
    "gadsden",
    "gilchrist",
    "glades",
    "gulf",
    "hamilton",
    "hardee",
    "hendry",
    "hernando",
    "highlands",
    "hillsborough",
    "holmes",
    "indian river",
    "jackson",
    "jefferson",
    "lafayette",
    "lake",
    "lee",
    "leon",
    "levy",
    "liberty",
    "madison",
    "manatee",
    "marion",
    "martin",
    "miami-dade",
    "monroe",
    "nassau",
    "okaloosa",
    "okeechobee",
    "orange",
    "osceola",
    "palm beach",
    "pasco",
    "pinellas",
    "polk",
    "putnam",
    "st. johns",
    "st. lucie",
    "santa rosa",
    "sarasota",
    "seminole",
    "sumter",
    "suwannee",
    "taylor",
    "union",
    "volusia",
    "wakulla",
    "walton",
    "washington",
]

fl_counties_pop_csv = [
    "saint johns",
    "palm beach",
    "duval",
    "marion",
    "lee",
    "broward",
    "pinellas",
    "volusia",
    "leon",
    "lake",
    "polk",
    "orange",
    "hillsborough",
    "charlotte",
    "putnam",
    "hernando",
    "okeechobee",
    "brevard",
    "levy",
    "sarasota",
    "de soto",
    "manatee",
    "collier",
    "seminole",
    "jackson",
    "pasco",
    "miami-dade",
    "martin",
    "saint lucie",
    "bay",
    "flagler",
    "alachua",
    "clay",
    "nassau",
    "escambia",
    "santa rosa",
    "walton",
    "okaloosa",
    "indian river",
    "citrus",
    "sumter",
    "osceola",
    "highlands",
    "taylor",
    "suwannee",
    "hendry",
    "monroe",
    "holmes",
    "wakulla",
    "columbia",
    "washington",
    "gadsden",
    "gulf",
    "hardee",
    "gilchrist",
    "calhoun",
    "baker",
    "bradford",
    "liberty",
    "dixie",
    "madison",
    "lafayette",
    "union",
    "franklin",
    "jefferson",
    "hamilton",
    "glades",
]

# print the misisng counties
missing_counties = set(fl_counties) - set(fl_counties_pop_csv)
print("Differing Counties:", missing_counties)
